fix: read dicom headers from template path, change dicom headers after processing, changing int8 to int16 so user can upload dicoms with smaller headers part, compress dicom headers error map withour offset

+++ ml_rdhei/backend/predictor/predict.py
@@ -42,7 +42,7 @@ def dicom_raw_ad_sklearn(batch: torch.Tensor, K: int = 5) -> Iterator[tuple[torc
 
     for img1, img2_X, img2_y, img2_ref_pixels in zip(img1_batch, X_img2_batch, y_img2_batch, ref_pixels_img2_batch):
         # image1 -> fixed prediction
-        img1_error_map = (15 - img1.flatten()).to(torch.int8)
+        img1_error_map = (15 - img1.flatten()).to(torch.int16)
         
         # image2 -> classic approach
         img2_kernel_weights, img2_error_map = sklearn_ridge(img2_X, img2_y)
@@ -56,7 +56,7 @@ def dicom_raw_ad_torch(batch: torch.Tensor, K: int = 5) -> tuple[torch.Tensor, t
 
     img1_batch, img2_batch = lr_decompose(batch)
 
-    img1_error_map_batch = (15 - img1_batch.flatten()).to(torch.int8)
+    img1_error_map_batch = (15 - img1_batch.flatten()).to(torch.int16)
 
     X_img2_batch, y_img2_batch, ref_pixels_img2_batch = extract_features(img2_batch, mask, K)
     kernel_weights_img2_batch, error_map_img2_batch = torch_ridge(X_img2_batch, y_img2_batch)
